+++ src/lesson35/dao/HotelDAO.java
@@ -33,7 +33,7 @@ public class HotelDAO extends GeneralDAO<Hotel> {
     public Hotel convert(String string) throws BadRequestException {
         String[] data = string.split(", ");
         if (data.length == 5) {
-            return new Hotel(Integer.parseInt(data[0]), data[1], data[2], data[3], data[4]);
+            return new Hotel(Long.parseLong(data[0]), data[1], data[2], data[3], data[4]);
         }
         throw new BadRequestException("Database error");
     }
