+++ src/lesson35/dao/UserDAO.java
@@ -37,7 +37,7 @@ public class UserDAO extends GeneralDAO<User> {
     public User convert(String string) throws BadRequestException {
         String[] data = string.split(", ");
         if (data.length == 5) {
-            return new User(Integer.parseInt(data[0]), data[1], data[2], data[3], UserType.valueOf(data[4]));
+            return new User(Long.parseLong(data[0]), data[1], data[2], data[3], UserType.valueOf(data[4]));
         }
         throw new BadRequestException("Database error");
     }
